+++ src/constants/carts.ts
@@ -1,0 +1,1 @@
+export const CART_LIMIT = 50;
