+++ src/pages/ProductsPage/ProductsPage.tsx
@@ -1,11 +1,9 @@
 import SelectBox from '../../components/common/SelectBox/SelectBox';
 import Header from '../../components/Header/Header';
-import { useEffect, useState } from 'react';
 import ProductList from '../../components/ProductList/ProductList';
 import useGetProducts from '../../hooks/useGetProducts';
 import { CATEGORY } from '../../constants/products';
 import useGetCarts from '../../hooks/useGetCartItems';
-import { AddCartItemType } from '../../types/cartItem';
 import Toast from '../../components/common/Toast/Toast';
 import {
   productPageContainer,
@@ -13,11 +11,10 @@ import {
   productPageTitle,
   selectBoxContainer,
 } from './ProductsPage.style';
-import postCartItem from '../../api/postCartItem';
-import deleteCartItem from '../../api/deleteCartItem';
 import useProductSort from '../../hooks/useProductSort';
 import { SORT } from '../../constants/products';
 import useProductCategory from '../../hooks/useProductCategory';
+import useCartManagement from '../../hooks/useCartManagement';
 
 function ProductsPage() {
   const { category, handleChangeCategory } = useProductCategory();
@@ -28,10 +25,15 @@ function ProductsPage() {
     products,
   } = useGetProducts({ category, sort });
   const { isLoading: isLoadingCarts, isError: isErrorCarts, carts, refetchCarts } = useGetCarts();
-  const [itemCount, setItemCount] = useState(0);
-  const [isErrorAddCardItem, setIsErrorAddCardItem] = useState(false);
-  const [isErrorDeleteCardItem, setIsErrorDeleteCardItem] = useState(false);
-  const [isOverItemCounts, setIsOverItemCounts] = useState(false);
+
+  const {
+    handleAddCartItem,
+    handleDeleteCartItem,
+    isErrorAddCardItem,
+    isErrorDeleteCardItem,
+    isOverItemCounts,
+    itemCount,
+  } = useCartManagement({ refetchCarts, carts });
 
   const getProcessedCartArr = () => {
     const cartIdArr = carts?.map((cart) => cart.product.id);
@@ -49,51 +51,6 @@ function ProductsPage() {
       };
     });
   };
-
-  const handleAddCartItem = async ({ productId, quantity }: AddCartItemType) => {
-    if (itemCount >= 50) {
-      setIsOverItemCounts(true);
-      setTimeout(() => {
-        setIsOverItemCounts(false);
-      }, 3000);
-
-      return;
-    }
-
-    const res = await postCartItem({
-      productId,
-      quantity,
-    });
-
-    if (!res.ok) {
-      setIsErrorAddCardItem(true);
-      setTimeout(() => {
-        setIsErrorAddCardItem(false);
-      }, 3000);
-    }
-
-    await refetchCarts();
-  };
-
-  const handleDeleteCartItem = async ({ productId }: { productId: number }) => {
-    const cartId = carts?.filter((cart) => cart.product.id === productId)[0].id || 0;
-    const res = await deleteCartItem({ cartId });
-
-    if (!res.ok) {
-      setIsErrorDeleteCardItem(true);
-      setTimeout(() => {
-        setIsErrorDeleteCardItem(false);
-      }, 3000);
-    }
-
-    await refetchCarts();
-  };
-
-  useEffect(() => {
-    if (carts) {
-      setItemCount(new Set(carts?.map((cart) => cart.product.id)).size);
-    }
-  }, [carts]);
 
   if (isLoadingCarts) {
     return <div>로딩중...</div>;
